Compute the melting behaviour of the crystals
This is an analysis of how the crystal structures evolve over time,
which has elucidated some incredibly interesting information that would
typically not be noticed by traditional crystal search methods. The
machine learning algorithm in looking for a number of different crystal
structures is able to accurately draw out additional information.