Create make analysis workflow
This is the creation of a workflow that allows for the calculation of
melting rates just with the command `make analysis`. This is comprised
of a python script, `melting_rates.py` which has been constructed from
work I have done in the jupyter notebooks. Additionally I have tied the
analysis to the data using a Makefile enabling simple and consistent
replication.