Remove code to calculate crystal parameters from notebook
This code to calculate the fraction of crystal is slow and so doesn't
belong in an interactive notebook. Additionally I the code can now be
run with `make analysis` which is a far more suitable interface for
this.

This just means the input data now has to be calculated outside the
notebook.