Only calculate crystal quantities every 100th step
Rather than calculate all the parameters for crystal growth on each
step, because these are time consuming calculations instead only look at
every 100th step by default.

This includes modifications to the script that runs the calculations,
setting the default values. In addition this is made explicit in the
makefile.