Compute melting rate normalised by the surface area
This is the culmination of a significant body of work in the calculation
of melting rates as a function of the surface area. Which is initially
proposed in logbook/2018-03-17.md in that assuming the interface is flat
is not a good assumption.

There is still simulations #13, #3 and analysis #9 to be conducted
however this fixes #8 and makes the previous issues solvable.