Change p2 cell size to 'magic' number
The p2 crystal doesn't have a orthorhombic unit cell. However, to make
everything easier I am making the unit cell orthorhombic, which turns
out to not work in the general case. It requires that the number of
layers lines up with the angle of the unit cell. A layer count that
seems to work is 42, which I had previously found, so I am using that
for all the crystals.