Use a lower temperature for creation of unit cells
There was an issue with the pg simulations which were breaking at the
first equilibration step being the melting of the outer interface. This
was carried through from the previous simulation as it occurred in the
first few timesteps of the melting simulation.

I have reduced the temperature of initialisation from 0.80 to 0.60 as a
fix to solve this issue.