Bugfixes: This fixes a number of issues with the simulations
There are a number of issues with the simulations I have conducted which
are documented in the file logbook/2018-04-09.md. With significant
changes to the simulations.

I have also taken the opportunity to move the directory melting long to
the appropriate date, since it appears I was still unsure of the year in
February.