Update environment.yml file
This updates the dependencies to ensure that running Conda env update
works. There appeared to be some issues with running pip and conda
together which occasionally resulted in errors when running the update.
This installs all the dependencies using conda, which solves the issue
somewhat, even if I don't particularly like the solution.