Track experiment.yml files in repository
This required a degree of gitignore trickery to get this to work
properly. Files in a gitignore file can only be unignored if their
parent folder is not ignored. Usually this would involve specifically
listing the parent directory, however this approach of using two levels
of wildcards in the ignore appears to be working.