ENV Include ipywidgets in env for better jupyter support
The widgets are really useful for building notebooks which are
relatively short yet are able to compare a whole range of different
values. There are times where I want to be able to look at all the
figures, yet others where I don't care about most of them and the
widgets are a good method of achieving this.