EXP Create an experiment for all interface data
This will collect all the data required for the interface simulations,
including both the low and high pressure simulations. The simulations
are run for long timescales to allow for good data collection.

There is possibly a requirement for multiple simulations for good data
collection, which is not yet included in this experiment.