ENH Create defects helper module
This is a module of helper functions for the analysis and creation of
crystal defects. These functions have been taken from
notebooks/06_Defect_Creation.ipynb and
notebooks/07_Defect_thermodynamics.ipynb and will be later imported back
into the notebooks to remove much of the code making them more readable.
Additionally in bringing the functions into a python module they are
easier to test and there is less reuse of code, with both notebooks now
able to use these functions.