ENV Create tmuxinator project
Adds a .tmuxinator.yml file to the project directory fixing #7. This
allows for a consistent project layout whenever working on the project.
Much like having a particular layout of equipment on the lab bench. By
having consistency it brings spatial awareness of where things are when
working on the project.