EXP Removing old/dated experiment directories
I have had a directory with the date of each experiment, however now that I am using
experi to manage the experiments and including the files in the git repo this practice
is somewhat redundant. Having all these direcotories and files also makes everything
more cluttered. I am now moving to a system where the only experiments that are retained
are a single instance of the experiment I used for data collection. This makes it much
simpler to replicate the results.