NB Update Crystal Detection and Clustering notebooks
This updates these notebooks bringing in the learnings from the
MLCrystals repository. The most notable change is the use of the
AgglomerativeClustering algorithm for the spatial clustering of
molecules.

Another change which significantly improves the readability of the
notebooks is the movement of much of the functions into the src
directory which are then imported. These functions detracted from the
story of the notebook and hence are not required.