MAINT Ensure all directories are part of the repository
The figures and models directories, which only contain files once
various components of the project are run should always exist in the
repository. This makes it much easier to program, meaning I can make the
assumption they are in place. This is done by adding a hidden .gitkeep
file in the directories so they are created upon the cloning of the
repository.