analysis: Update Makefile with new dynamics analysis
The introduction of the dynamics_calc module means that the analysis
procedure for calculating the dynamics properties is much more in depth.
This commit introduces a collection of new Makefile rules to handle the
complexity of this updated procedure. There are also a few variables
modified in the melting analysis to prevent variable overlap with the
dynamics.