exp: Create experiment to get energy of crystals
Part of the development of the crystal melting behaviour requires
knowing the energy difference between the liquid and crystal
configurations at a specific temperature. I am able to get the data for
the liquid from the dynamics calculations, however I need to calculate
the value for the crystal, which is what this set of experiments is for.