analysis: Create tool for analysing melting rates
This creates a tool for the calculation of melting rates from the
melting data. As part of this calculation there is also the inclusion of
a tool for the cleanup of the data.

These tools have been moved from the jupyter notebooks making them
easier to use and to incorporate them into the dependencies of the
makefile.