nb: Simplify calculations of melting rates
This makes the notebook containing the plotting of the melting rates
significantly simpler. The previous set of commits has been moving the
functionality out of the notebook and into scripts, so this commit is
making use of all those changes and significantly simplifying the
plotting and analysis of the melting rates.