doc: Add README files to each experiment
Having a README file documenting the aims of each of the experiments and
why they are being conducted. This is to help motivate each of them and
keep them nice and contained. Additionally as there are now lots of
experiments and more coming it is much more likely to get really
confused about the goals of each of them.