analysis(melting): Update melting rates notebook
This makes a series of changes to the melting rates notebook which
includes bypassing a bug in pandas, where the read_hdf doesn't close a
file when the key doesn't exist.

This also now uses the latest data from the melting rates experiment,
saving the files for inclusion in the repository and the thesis.