analysis(melting): Reduce minimum volume in cleanup
With the change to the analysis that I am taking into account the
changing surface area, this reduces the minimum volume for the cleanup.
It is important to still have a minimum volume to deal with any unusual
behaviour at small volumes, with the fluctuations being much larger
relative to the overall size.