analysis: Split analysis of melting behaviour and rates
This splits the makefile analysis of the melting, which is all about the
melting behaviour, in particular the solid state phase transition parts
along with the stability of the p2 crystal over the other crystals. The
other component is the rate, where the melting rate of the p2 crystal is
calculated.