env: Pin all dependencies
This pins the version number of all the dependencies, not just the
versions of my analysis simulation programs. This pinning will ensure
that everything updates when the environment is updated and there should
be fewer issues overall. Currently the versions are the latest versions
of everything, which can be reassessed at later points in time when
everything updates.