chore: Use markdown for notebooks
This uses a combination of markdown and the jupytext package for
handling the notebook files. Each markdown file is linked to a notebook
file which is not included in the repository.

The use of markdown offers a range of advantages, the most notable being
a simplicity of working with the files, seeing the changes which have
been made in diffs and searching through the files. The markdown files
are also much simpler to work with in a text editor which is where a lot
of the development will be taking place.