analysis: Create rule to sync notebooks and Markdown
I am storing Markdown files in the repository, however I interact with
the jupyter notebooks in the browser. So this adds a new command to the
makefile which ensures that both the notebooks and the Markdown files
are in sync, additionally adding the correct configuration such that
they should remain in sync.