experiment(thermodynamics): Match dynamics temperatures
The thermodynamics experiment has previously been getting the free
energy of the crystal. However, in #33, we now want an analysis of how
the free energy changes, in addition to an estimation of the surface
free energy at a range of temperatures.

This generates the crystal structures up to the spinodal point for both
low and high pressure, matching the values for the dynamics, since the
dynamics results are valid for the liquid configuration.