analysis: Create function to normalise temperature
The normalised temperature is a useful metric, and I have now saved the
melting points in a file. This creates a function to read in the
results/melting_points.csv file, and use the contained values as the
melting points.

This makes it easy to add additional parameters and importantly update
the best estimate of the melting point in a simple way.

fixes #34