notebook: Update melting classification with new temps
The interface experiment has been modified to include a much smaller
number of temperatures. This updates the melting classification notebook
to use these temperatures, and additionally updating the timesteps at
which the interesting things happen within a configuration.

As part of this update, all of the relevant figures have also been
updated, removing the old figures with the incorrect information in
their name and replacing them with the new figures.