feat: Use traj3dy utility for some analysis
I have written a utility in rust, which offers the performance of C++
with a composability similar to that of python, particularly for dealing
with a command line interface. This uses traj3dy deployed using conda,
for a smattering of the analysis, typically leaving the existing
analysis in place.

This incorporates the fluctuation analysis in the repository, fixing #33