fix: Include calculation of enthalpy
This doesn't just calculate the change in the potential energy, there is
additionally the enthalpy which is the change of the entire system. This
is the quantity which will be used to calculate the melting rates.

This moves the calculation of the resulting values to the enthalpy file
in the results folder, and includes the summary values.

Ref #49