exp(disc): Update disc experiments for correct MP
This modifies the set of temperatures which the simulations are running
at to be around the correct melting point for the 2D LJ crystal, which
is 0.53.

In addition to the melting point changes, the simulations have been
simplified significantly making it easier to run and get results from
them.

Fixes #53