notebooks: Rewriting of melting analysis
This is a fairly extensive rewrite of the melting analysis, now
including the Turnbull model and adding other models of growth.

This also now properly includes the chemical potential energy, not just
using the same value, which has resulted in a number of changes to
figures and the rate values.