notebooks(spatial): Diverging figures
This implements a diverging colourscheme for the spatial dynamics
notebooks, with the central value being the mean relaxation time in the
liquid. This means I can easily visualise the molecules which are faster
and slower than the mean.

This also includes a significant usability update, storing the dynamics
analysis in a file which can be read in. There is the code to create the
file should it not already exist.